chore(server-split): v0.4.x PR-H — final extraction (5 router modules) + server.py 정리 (892 lines)
docs/design/v0.4.x-server-split.md 8-PR sequence 마지막 PR. 잔여 17 endpoints
+ 5 Pydantic 모델 → 5 새 router 모듈. server_llmwiki.py 정리.

## 이동된 surface (~600 lines from server)

### 5 new router modules

| 모듈 | endpoints | Pydantic | 목적 |
|---|---|---|---|
| routes/query.py     | /query/ (151L)                                  | QueryRequest, QueryResponse | 메인 질의 |
| routes/history.py   | /history/* × 6                                  | —                           | 대화 히스토리 |
| routes/feedback.py  | /feedback/, /feedback/stats/                    | FeedbackRequest             | 피드백 |
| routes/multimodal.py| /analyze/image, /analyze/video, /screen/analyze | ScreenRequest               | 멀티모달 |
| routes/ops.py       | /status/, /hardware/, /trace/poll/, /export/, / | StatusResponse              | 운영 |

### server_llmwiki.py 최종 형태

```
1. imports + utf-8 console
2. config + RateLimiter
3. _init_audit_db
4. app = FastAPI(...)
5. /static mount
6. middleware (no_cache_static, rate_limit_middleware)
7. 5 HTML shells (/, /healthz, /admin, /workspace, /admin/graph)
8. DI singletons (rag_engine, file_processor, _rate_limiter)
9. set_* registration + include_router × 12 (auth/llm/jobs/artifacts/evolution/coding/admin/query/history/feedback/multimodal/ops)
10. @app.on_event("startup") (135L plugin pack loader)
11. if __name__ == "__main__"
```

target ~500 lines (design memo). 실측 892 — 100% 가능했으나 on_startup
135L 가 단일 함수라 분할 안 함 + audit_db init + RateLimiter class 가
server-level concern 으로 유지. 5894 → 892 = **85% 감소**.

## 패턴: _LazySingleton (routes/query, history, feedback, multimodal, ops)

routes/admin.py 에서 도입된 `_LazySingleton` descriptor — 핸들러
`rag_engine.foo` / `file_processor.foo` syntax 보존하면서 routes/_deps
의 lazy-forwarder 통해 monkeypatch 호환. 5 새 모듈 모두 동일 패턴 채택.

## server back-compat re-exports

테스트가 server module attribute 로 접근하는 8개 이름 (#noqa: F401):
- bearer_scheme, verify_api_key, resolve_api_key_principal, get_role_from_request
- QueryRequest, QueryResponse, FeedbackRequest, ScreenRequest, StatusResponse

## tests 갱신 (9 파일 + helper)

- tests/_server_split_helpers.py: 12개 routes/* 모듈 모두 포함
- test_a2_phase2 / document_export / force_web_chip / observability / query_include_contexts / real_reasoning_stream / web_used_badge: combined_server_source() 사용

## 검증

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ 125 endpoints identical to origin/main |
| 2-5. middleware/audit/static/startup | ✓ smoke pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** |
| ruff F-class | ✓ All checks passed |

## 누적 8-PR 결과

```
server_llmwiki.py:  5894 → 892 lines (-5002, -85%)
```

| PR | scope | -lines |
|---|---|---|
| #572 PR-A.0 | foundation + audit gate + boot smoke | (+711) |
| #573 PR-A.1 | 13 auth/upload | -724 |
| #574 PR-B   | 12 LLM/Ollama | -485 |
| #575 PR-C   | 9 jobs + DI lazy | -246 |
| #576 PR-D   | 5 W7-A artifacts | -158 |
| #577 PR-E   | 13 evo/learn/patch | -589 |
| #578 PR-F   | 4 coding | -155 |
| #579 PR-G   | 47 admin + _LazySingleton | -2002 |
| #이번 PR-H  | 17 query/history/feedback/multimodal/ops | -643 |

## Out of scope (별 cycle)
- on_startup 135L 함수 분할 (별 PR)
- RateLimiter / _init_audit_db 를 core/ 로 추출 (별 PR)
- HTML shell route 통합 (현재 5개 서로 다른 위치)

Quality delta: exempt (label: chore). URL byte-identical, core/ 무관.
