+++ routes/multimodal.py
@@ -1,0 +1,136 @@
+"""Multimodal /analyze/* + /screen/analyze/ routes (PR-H)."""
+from __future__ import annotations
+
+import os
+import time
+from datetime import datetime
+from typing import Optional
+
+from fastapi import APIRouter, Depends, File, Form, HTTPException, UploadFile
+from pydantic import BaseModel
+
+from config import UPLOAD_DIR
+from routes._deps import get_file_processor, get_rag_engine
+from routes._helpers import (
+    _require_feature,
+    _write_audit,
+    get_role_from_request,
+    verify_api_key,
+)
+
+
+class _LazySingleton:
+    def __init__(self, getter):
+        object.__setattr__(self, "_getter", getter)
+    def __getattr__(self, name):
+        return getattr(self._getter(), name)
+
+
+rag_engine = _LazySingleton(get_rag_engine)
+file_processor = _LazySingleton(get_file_processor)
+router = APIRouter()
+
+# ─── Pydantic ───
+
+class ScreenRequest(BaseModel):
+    api_key:  str
+    question: str = ""
+    region:   Optional[list] = None   # [x, y, w, h]
+
+# ─── Endpoints ───
+
+@router.post("/analyze/image/", summary="이미지 분석 [P7-VIS-1]")
+async def analyze_image(
+    file:    UploadFile = File(...),
+    api_key: str = Form(...),
+    role:    str = Depends(get_role_from_request),
+):
+    """이미지 파일 업로드 → EXIF + LLaVA 분석 → 결과 반환."""
+    verify_api_key(api_key)
+    suffix  = os.path.splitext(file.filename)[1].lower()
+    allowed = {".jpg", ".jpeg", ".png", ".gif", ".webp", ".bmp"}
+    if suffix not in allowed:
+        raise HTTPException(status_code=400, detail=f"지원 형식: {allowed}")
+
+    # 임시 저장
+    tmp_path = os.path.join(UPLOAD_DIR, f"vis_{int(time.time())}{suffix}")
+    with open(tmp_path, "wb") as f:
+        f.write(await file.read())
+
+    try:
+        from tools.multimodal.image_analyzer import analyze_image as _analyze
+        result = _analyze(tmp_path)
+        _write_audit(role, "/analyze/image/", query=file.filename,
+                     answer=str(result.get("description",""))[:80])
+        return {
+            "filename":    file.filename,
+            "analyzed_at": datetime.now().isoformat(),
+            "description": result.get("description",""),
+            "date":        result.get("date",""),
+            "location":    result.get("location",""),
+            "persons":     result.get("persons",[]),
+            "tags":        result.get("tags",[]),
+            "exif":        result.get("exif",{}),
+            "success":     result.get("success", True),
+        }
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+    finally:
+        try: os.remove(tmp_path)
+        except Exception: pass
+
+@router.post("/analyze/video/", summary="영상 분석 [P7-VID-1]")
+async def analyze_video(
+    file:    UploadFile = File(...),
+    api_key: str = Form(...),
+    role:    str = Depends(get_role_from_request),
+):
+    """영상 파일 업로드 → OpenCV 장면 + Whisper 자막 분석."""
+    verify_api_key(api_key)
+    suffix  = os.path.splitext(file.filename)[1].lower()
+    allowed = {".mp4", ".avi", ".mov", ".mkv", ".webm"}
+    if suffix not in allowed:
+        raise HTTPException(status_code=400, detail=f"지원 형식: {allowed}")
+
+    tmp_path = os.path.join(UPLOAD_DIR, f"vid_{int(time.time())}{suffix}")
+    with open(tmp_path, "wb") as f:
+        f.write(await file.read())
+
+    try:
+        from tools.multimodal.video_analyzer import analyze_video as _analyze
+        result = _analyze(tmp_path)
+        _write_audit(role, "/analyze/video/", query=file.filename,
+                     answer=str(result.get("summary",""))[:80])
+        return {
+            "filename":    file.filename,
+            "analyzed_at": datetime.now().isoformat(),
+            "summary":     result.get("summary",""),
+            "duration":    result.get("duration",""),
+            "scenes":      result.get("scenes",[]),
+            "transcript":  result.get("transcript",""),
+            "tags":        result.get("tags",[]),
+            "success":     result.get("success", True),
+        }
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+    finally:
+        try: os.remove(tmp_path)
+        except Exception: pass
+
+@router.post("/screen/analyze/", summary="화면 분석 [P7-SCR-1]")
+async def screen_analyze(
+    data: ScreenRequest,
+    role: str = Depends(get_role_from_request),
+):
+    """화면 캡처 → OCR → LLM 분석. admin 전용."""
+    _require_feature(data.api_key, role, "admin.tools")
+    try:
+        from tools.screen.screen_agent import run_screen_analysis
+        region = tuple(data.region) if data.region else None
+        result = run_screen_analysis(data.question, region)
+        _write_audit(role, "/screen/analyze/",
+                     query=data.question[:60],
+                     answer=result.get("analysis","")[:80])
+        return result
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
