+++ tests/test_web_used_badge.py
@@ -23,7 +23,6 @@ Run:
 """
 from __future__ import annotations
 
-import inspect
 import os
 import re
 import sys
@@ -91,8 +90,8 @@ class PipelineReturnShapeTests(unittest.TestCase):
 class QueryResponseShapeTests(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        import server_llmwiki as srv
-        cls.src = inspect.getsource(srv)
+        from tests._server_split_helpers import combined_server_source
+        cls.src = combined_server_source()
 
     def test_response_model_has_fields(self):
         m = re.search(
